Medications api calls moved to separate hook
